requirements.txt
fastapi
uvicorn[standard]
azure-cosmos
python-dotenv
pydantic
pytest
httpx